[Chore] Move LogProb to a separate file
